chatbot/backend/requirements.txt
annotated-types==0.7.0
attrs==25.4.0
aws-cdk-lib==2.222.0
aws-cdk.asset-awscli-v1==2.2.242
aws-cdk.asset-node-proxy-agent-v6==2.1.0
aws-cdk.cloud-assembly-schema==48.18.0
aws_lambda_powertools==3.23.0
boto3==1.40.64
botocore==1.40.69
cattrs==25.3.0
constructs==10.4.3
importlib_resources==6.5.2
jmespath==1.0.1
jsii==1.118.0
publication==0.0.3
pydantic==2.12.3
pydantic_core==2.41.4
python-dateutil==2.9.0.post0
s3transfer==0.14.0
six==1.17.0
typeguard==4.2.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
